refactor account details to new tab layout (first pass)
