+++ frontend/src/components/tabs/StyledTabs.module.scss
@@ -1,0 +1,27 @@
+@import "styles/spacings";
+@import "styles/colors";
+
+.tab {
+  max-width: 120px;
+  padding: $spacing-s !important;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.activeTab {
+  background: $gray-100 !important;
+  border-bottom: 3px solid $violet-100 !important;
+}
+
+.tabLabel {
+  text-align: center;
+}
+
+.tabWrapper, .inactiveTab {
+  background: $gray-110;
+}
+
+.content {
+  padding: $spacing-base 0;
+}
