+++ frontend/src/pages/accounts/details/storage/AccountStorage.module.scss
@@ -1,0 +1,49 @@
+.grid {
+  width: 100%;
+  max-width: 100%;
+  display: grid;
+  column-gap: 15px;
+  row-gap: 20px;
+  grid-template-columns: repeat(3, 1fr);
+  margin-bottom: 40px;
+}
+
+.gridExtendable {
+  width: 100%;
+  max-width: 100%;
+  display: grid;
+  column-gap: 15px;
+  row-gap: 20px;
+  grid-template-columns: repeat(4, 1fr);
+
+  .gridItemExtended {
+    grid-column: 1 / span 4;
+    height: 600px;
+  }
+}
+
+@media all and (max-width: 1350px) {
+  .gridExtendable {
+    grid-template-columns: repeat(3, 1fr);
+
+    .gridItemExtended {
+      grid-column: 1 / span 3;
+      height: 600px;
+    }
+  }
+}
+
+@media all and (max-width: 900px) {
+  .grid {
+    grid-template-columns: repeat(2, 1fr);
+  }
+
+  .gridExtendable {
+    grid-template-columns: repeat(2, 1fr);
+
+    .gridItemExtended {
+      grid-column: 1 / span 2;
+      height: 600px;
+    }
+  }
+}
