+++ frontend/src/pages/accounts/details/storage/cards/BaseStorageCard.module.scss
