+++ frontend/src/pages/accounts/details/storage/cards/InternalStorageCard.tsx
@@ -3,9 +3,9 @@ import classes from "./BaseStorageCard.module.scss";
 import { AccountStorageItem } from "@flowser/shared/dist/src/generated/entities/accounts";
 import { FlowUtils } from "utils/flow-utils";
 import classNames from "classnames";
-import { SimpleButton } from "../../../components/buttons/simple-button/SimpleButton";
+import { SimpleButton } from "../../../../../components/buttons/simple-button/SimpleButton";
 import { DecoratedPollingEntity } from "contexts/timeout-polling.context";
-import { JsonView } from "../../../components/json-view/JsonView";
+import { JsonView } from "../../../../../components/json-view/JsonView";
 
 type ExtendableStorageCardProps = {
   storageItem: DecoratedPollingEntity<AccountStorageItem>;
