+++ frontend/src/pages/accounts/details/storage/cards/PublicPrivateStorageCard.module.scss
