+++ frontend/src/pages/interactions/components/templates/InteractionTemplates.module.scss
@@ -8,7 +8,6 @@
   padding: $spacing-l;
   display: flex;
   flex-direction: column;
-  height: 100%;
   justify-content: space-between;
 
   .storedTemplates {
